refactor namespaces of tests to match the package change in the classes being tested
